troubleshoot cont
Next idea: change the data used in the catch likelihood, currently it is using data from bethel, this may be too small of a number, might be more appropriate to use commercial and subsitence catch. changing that now -- didnt matter, that is the same as the bethel effort. expanding date range to match excel sheet to see if I can get predicted catch to match exactly at least...