commit before adding environmental covariates
spent a long time messing around with priors. I got all parametrs to align, except C (carrying capacity) which is notoriously hard to estimate? so for now goign to move on to adding in the environmental info, but can return to thsi push if it turns out i needed C to be exact on this stage.