it works
but I am getting bad estimates, made priors smaller to see if that helps now that i found error. can run with initial values too, but I don't think that'll matter toooo much.
The issue was that q=n_returning/sum(n_returning) was encountering the NAs and getting an error, so I changed the index to not index the incomplete return years, and that seems to have fixed it.
Next - add a covariate back in.