fishin prior
sensitive to fishing mortality prior, when I change it from (0,10) to (1,2) it gets mad
 - trying to change mean but keep variance- normal (1,10) --> NOPE
normal (-0.5,10) --> works
but seems like it really overestimates the peak in fishing at time point 5, there is an icnrease but its going super high
changes prior to normal (-0.5, 5) and (-0.5,3) and it runs, but at this point it doesnt change estiamtes. Fishing variance is still really large.