cont working on mod
added extension of years so it estimates fully up to all the brood years supplied in basis and captures the recent drop off
trying to estimate more starting value to help with the model running but no luck so far.
- adding a lower bound to the starting value is the key!! after that is added it works...
currently model captures patterns but is way more variable than data, this pattern changes with changes in starting values, so i think that is why they need to be estimated....

changed covariate priors to (0,1) isntead of (0,10) - the model runs.

Fishing mortality: truncated normal distribution helps the model revert to a certain value with out information, trying a normal(0,0.1) prior, it wont run if 0 is the mean, it wont run if -0.05 is the mean
-  it will run with -1 as the mean.
- will NOT work with -0.9...