fixing sigma
good to fix for catch and recruits, when i fix it for spawners things get kinda weird.

seeing what it looks like when i fix for juveniles.

fixing FM to v low fishign after time step 20 seems to work well...  not sure if this is best though. need to check wih curry to find better solutions.