marine mortality
something is weird here. but i cant fifgure out why, how is marine mortality decreasing, when spawners and juveniles decrease. it seems like the pattern i am seeing, hi marine mortality then low marine mortaliyt should be reverse. the juv and marien survival rates are very different and i am not sure if they should be.

- One thing that is weird is that c1 is estimating lower than c2 and with little SD, red flag.
- Fits adult estimates pretty well but juvenile estimates are off.

- Worse when sigma j is fixed at a low CV, fixing it at a higher CV to see what happens there. Maybe best to just to keep estimating it.
    - CV from model I also don’t think was great
- Worse when basal p is tied or constrained.
- Need to do a deeper dive into marine and juvenile survival because it seems off still.
- Theta 1s have a big CI and theta2 have a small lCI what’s up with that. Something is weird between these for sure….
- P1 hits 1 and then p2 is super small….

going to run simulation and see what its like there ....

notes on sim:

isnt estimating theta accurately, all are identical.  decreased prior IDs, main take home is that in the simulation the kappas are not as volatile as in the actual mode....

- paying attention to how changing basal survival changes model
- basal at 0.01 and 0.1, more realistic, I think i was setting the basal ps too high.
- trying it at 0.1 and 0.3. - much more variation here in kappa survival and mortlaity
- trying it at 0.08 and 0.2 this looked better, running data with this as startign values.