messing with Fm, addint penalty
added a penalty on the fishing deviation to try and make it not go so crazy in recent years.

fixed the deviations to (0,1) so that it is penalzied to go to 0 isntead of estiamte huge deviations.

now it isnt fitting age comps correctly. met wiht kimberly going to try trouble shooting.