meeting with curry
suggests the problems i am havign will be fixed if i loosen CV on the juvenile data, basically its the dataset with the least amount of certainty so its ok if we don't make it fit it completely.

estimating with one covariate and survival on both stages.
sensitive to addition of covariates. NExt switching around brood year indexing it may be off by 1 year??