juv index/change indexing
2020 juv index had mean in it, but it shouldn't have been that high, instead changed it to a rolling mean with the mean of surrounding 5 years.

plotting to see if the juvenile index aligns -- it is two years off.
I see that its part of my indexing. goi to try and fix the indexing, and see if they align...

I changed the indexing, but TBH not it made much of a difference?? I think they should be t+a+2, but the fits don't get all that much better...added some more plots for comparisons...