simplifying model again, cont to fix indexing....
removing covariate effects when checking the indexing to make sure survival estimates ok without covariates.
Curry said indexing should be t+a+1 so changing back to this...
but currently cant get it to run!!

reverting back to a script i know runs, t+a+2 and trying that without the covariates,
Model is having issues running when it has to estimating p_1 and p_2 and not use covariates....
currently commiting as t+a+2, estiamting p_1 and having covariates for p_2... the p_1 s are crazy high though.