Time varying is IT
"basically from what i understand, the brood year variation in age is a hyperdistribution (dirichlet) arising from some  mean age at maturity vector. and then the multinomial is used for fitting the comp data" -- "thats one of the issues is because the alpha_sum controls how much p[t,a] can vary from year to year. so if there is not variability just based on how its parameterized, then g or alpha are not going to converge?"