re-running with a looser theta prior - all good
moved from 0.001 to 1 just to see if it is still estimable.

aside from that, all looks decent, age comp trace plots still kinda funky. gonna hold this simulation as is for matt on monday and re-set up the data portion with some updates and see how that goes.

it still found theta with a sigma prior of  1, so thats good. had a few more warnings though, so it seems like it preferes the tighter prior, harvest is weird in a few spots, so may need to revert to the deviation method, lets see how it fits with data.