cont. its working ok
previous run that worked well didn't evaluate the likelihood till time step 6, i just changed it to start at 1 to see if that changes the wonky starting values that are estimated.

working ok, estiamted starting values are kinda wonky when sigma is estimated.
Going to test a few ways to decrease sigma for spawners/returns and/or starting values.