remove sigma catch
if catch is fixed, what happens - everything else gets flatter.  not the move.

loosened up the sigma on catch, but is still fixed.
things get a little better, but the juvenile time series and survival isn't great.

fix basal 2 and estimate basal 1! p1 is flat which i think is ok if the covariate effect is 0?
- if i add basal_p_1, it estimates spuepr low, but it doesn't change the j survival or productivity.
- reversed: same, juvenile productivity is still flat.

next: add more covariates.