curry sent an updated juvenile index. update with new juv data!
updated with new juv data and ran model. This also added a couple new years of data.
Currently I am using mean annual proportion of fall chum in the categories, and sabrina is going to model it eventually? the index is different when I do this, because each strata was surveyed for genetics at different amounts...may need to check with curry and revisit this at some point...

outcomes: some divergent transitions (0 before with old index)