likelihood structure back to normal.
estimating eggs using a ricker transition... i am hoping this allows better fits to the juvenile estimates and allows more flexibility in the model. but it requires two more parameters that are not informed by data....
got it to run with alpha beta fixed, next try and estimate them.

estimate alpha and beta - improves estimates like i was hoping... I think i still need to change the accounting in the model though so it doesn't do it for every age?

adjust hatchery covariate - but not running until i address the stomach fullness issues with curry